chore: installed d3 and d3 for typescript
