+++ client/vite.config.ts
@@ -4,6 +4,9 @@ import react from "@vitejs/plugin-react";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  optimizeDeps: {
+    include: ["d3"]
+  },
   server: {
     proxy: {
       "/api": {
